Medium-Clone/backend/requirements.txt
Django==4.2.7
djangorestframework==3.14.0
django-cors-headers==4.3.1
drf-spectacular==0.26.5
python-decouple==3.8
psycopg2-binary==2.9.9
Pillow==10.1.0
django-imagekit==4.1.0
redis==5.0.1
django-redis==5.4.0
celery==5.3.4
django-celery-beat==2.5.0
django-celery-results==2.5.0
channels==4.0.0
channels-redis==4.1.0